Separate UserData from contract to be reused. Separate OnBoardingLoading-state to LoadingState-enum to be reused.
