+++ app/src/main/java/com/antisnusbolaget/slutasnusa2/viewmodel/LoadingState.kt
@@ -1,0 +1,7 @@
+package com.antisnusbolaget.slutasnusa2.viewmodel
+
+enum class LoadingState {
+    LOADING,
+    FAILED,
+    SUCCESS,
+}
