+++ app/src/main/java/com/antisnusbolaget/slutasnusa2/viewmodel/UserData.kt
@@ -1,0 +1,7 @@
+package com.antisnusbolaget.slutasnusa2.viewmodel
+
+data class UserData(
+    val costPerUnit: Int,
+    val units: Int,
+    val dateWhenQuit: Long,
+)
