+++ app/src/main/java/com/antisnusbolaget/slutasnusa2/viewmodel/homescreen/HomeScreenContract.kt
@@ -1,0 +1,9 @@
+package com.antisnusbolaget.slutasnusa2.viewmodel.homescreen
+
+import com.antisnusbolaget.slutasnusa2.viewmodel.LoadingState
+
+sealed interface HomeScreenEvent
+
+data class HomeState(
+    val loadingState: LoadingState = LoadingState.LOADING,
+)
