+++ app/src/main/java/com/antisnusbolaget/slutasnusa2/viewmodel/homescreen/HomeScreenViewModel.kt
@@ -1,35 +1,32 @@
-package com.antisnusbolaget.slutasnusa2.viewmodel
+package com.antisnusbolaget.slutasnusa2.viewmodel.homescreen
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.antisnusbolaget.slutasnusa2.data.DataStoreRepo
+import com.antisnusbolaget.slutasnusa2.viewmodel.LoadingState
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 import timber.log.Timber
-
-sealed interface HomeScreenLoadingState {
-    object Loading : HomeScreenLoadingState
-    object Failed : HomeScreenLoadingState
-    object Success : HomeScreenLoadingState
-}
 
 class HomeScreenViewModel(
     dataStoreRepo: DataStoreRepo,
 ) : ViewModel() {
 
-    private val _uiState: MutableStateFlow<HomeScreenLoadingState> = MutableStateFlow(HomeScreenLoadingState.Loading)
-    val uiState: StateFlow<HomeScreenLoadingState> = _uiState
+    private val _uiState =
+        MutableStateFlow<HomeState>(HomeState())
+    val uiState: StateFlow<HomeState> = _uiState
 
     init {
         viewModelScope.launch {
             dataStoreRepo.isKeyStored().collect {
                 if (it) {
                     Timber.d("Osk, is key stored? $it")
-                    _uiState.value = HomeScreenLoadingState.Success
+                    _uiState.update { uiState.value.copy(loadingState = LoadingState.SUCCESS) }
                 } else {
                     Timber.d("Osk, is key stored? $it")
-                    _uiState.value = HomeScreenLoadingState.Failed
+                    _uiState.update { uiState.value.copy(loadingState = LoadingState.FAILED) }
                 }
             }
         }
